# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Rentway API'
  description: 'Mobile-friendly REST API v1. Secured endpoints require a Bearer token obtained via /api/v1/auth/login. API Owner: Sudharma HewaVitharana.'
  version: 1.0.0
servers:
  -
    url: /api
    description: 'API base path'
paths:
  /v1/ads:
    get:
      tags:
        - Ads
      summary: 'List ads'
      operationId: 9b15c4b7a7b6fd8d42d406fc790eaa46
      parameters:
        -
          name: company_id
          in: query
          required: false
          schema:
            type: integer
        -
          name: status
          in: query
          required: false
          schema:
            type: string
        -
          name: per_page
          in: query
          required: false
          schema:
            type: integer
      responses:
        200:
          description: OK
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Ads
      summary: 'Create ad'
      operationId: 9a15268410d57e0bd72b4deb6f520a56
      responses:
        201:
          description: Created
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        422:
          $ref: '#/components/responses/ValidationError'
      security:
        -
          bearerAuth: []
  '/v1/ads/{id}':
    get:
      tags:
        - Ads
      summary: 'Show ad'
      operationId: 6c738b9e04df04c37db844adbc36e9e7
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Ads
      summary: 'Update ad'
      operationId: b0f18b25b859468c9b7e3a092deff3fe
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Updated
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        422:
          $ref: '#/components/responses/ValidationError'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Ads
      summary: 'Delete ad'
      operationId: 71f7998c0cda61119b940c108bccf865
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Deleted
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
      security:
        -
          bearerAuth: []
...
  /v1/auth/register:
    post:
      tags:
        - Auth
      summary: Register
      description: 'Register a new user.'
      operationId: 58fb87dce972174198f99707c002431d
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  example: 'Jane Doe'
                email:
                  type: string
                  example: jane@example.com
                password:
                  type: string
                  example: secret123
                usertype:
                  type: string
                  example: customer
              type: object
      responses:
        201:
          description: Registered
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Registered
                  data:
                    properties:
                      token:
                        type: string
                      user:
                        properties:
                          id:
                            type: integer
                            example: 1
                          name:
                            type: string
                            example: 'Jane Doe'
                          email:
                            type: string
                            example: jane@example.com
                        type: object
                    type: object
                type: object
  /v1/auth/login:
    post:
      tags:
        - Auth
      summary: Login
      description: 'Login with credentials.'
      operationId: 3ff5fdf1e87b064055e72ef143ecc7c5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: jane@example.com
                password:
                  type: string
                  example: secret123
              type: object
      responses:
        200:
          description: 'Logged in'
        401:
          description: 'Invalid credentials'
  /v1/auth/logout:
    post:
      tags:
        - Auth
      summary: Logout
      description: 'Logout current user (invalidate tokens).'
      operationId: 5df06d8bf07ab28923f168580a1b2323
      responses:
        200:
          description: 'Logged out'
      security:
        -
          bearerAuth: []
  /v1/auth/me:
    get:
      tags:
        - Auth
      summary: Me
      description: 'Get authenticated user.'
      operationId: 43b10aadfac909e8840e7bd830e76590
      responses:
        200:
          description: 'User data'
      security:
        -
          bearerAuth: []
...
  /v1/expenses:
    get:
      tags:
        - Expenses
      summary: 'List expenses'
      operationId: e72f19fed0591b9616f150c948c156e4
      parameters:
        -
          name: company_id
          in: query
          required: false
          schema:
            type: integer
        -
          name: expenses_type
          in: query
          required: false
          schema:
            type: string
        -
          name: per_page
          in: query
          required: false
          schema:
            type: integer
      responses:
        200:
          description: OK
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Expenses
      summary: 'Create expense'
      operationId: 0259d9961921405af70ce1f6b76c288e
      responses:
        201:
          description: Created
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        422:
          $ref: '#/components/responses/ValidationError'
      security:
        -
          bearerAuth: []
  '/v1/expenses/{id}':
    get:
      tags:
        - Expenses
      summary: 'Show expense'
      operationId: aafd70e4a772c7caa1948018d90d121a
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Expenses
      summary: 'Update expense'
      operationId: 27d77d3386ce32e0c69987440c316d85
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Updated
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        422:
          $ref: '#/components/responses/ValidationError'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Expenses
      summary: 'Delete expense'
      operationId: 4caa19ba86d6792735a9431592cef7c3
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Deleted
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
      security:
        -
          bearerAuth: []
...
components:
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Unauthorized
            type: object
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Forbidden
            type: object
    ValidationError:
      description: 'Validation failed'
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 'Validation failed'
              errors:
                type: object
            type: object
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...